+++ Admin/src/views/admin/default/index.jsx
@@ -13,7 +13,7 @@ import {
 // Assets
 import Usa from "assets/img/dashboards/usa.png";
 // Custom components
-// import MiniCalendar from "components/calendar/MiniCalendar";
+import MiniCalendar from "components/calendar/MiniCalendar";
 import MiniStatistics from "components/card/MiniStatistics";
 import IconBox from "components/icons/IconBox";
 import React from "react";
@@ -141,10 +141,10 @@ export default function UserReports() {
           columnsData={columnsDataComplex}
           tableData={tableDataComplex}
         />
-        {/* <SimpleGrid columns={{ base: 1, md: 2, xl: 2 }} gap='20px'>
+        <SimpleGrid columns={{ base: 1, md: 2, xl: 2 }} gap='20px'>
           <Tasks />
           <MiniCalendar h='100%' minW='100%' selectRange={false} />
-        </SimpleGrid> */}
+        </SimpleGrid>
       </SimpleGrid>
     </Box>
   );